checkstyle fix, change supports method at Command classes
