+++ bot/src/main/java/edu/java/bot/Bot.java
@@ -4,7 +4,6 @@ import com.pengrad.telegrambot.UpdatesListener;
 import com.pengrad.telegrambot.model.Update;
 import com.pengrad.telegrambot.request.BaseRequest;
 import com.pengrad.telegrambot.response.BaseResponse;
-
 import java.util.List;
 
 public interface Bot extends AutoCloseable, UpdatesListener {
