+++ bot/src/main/java/edu/java/bot/commands/Command.java
@@ -1,9 +1,8 @@
-package edu.java.bot;
+package edu.java.bot.commands;
 
 import com.pengrad.telegrambot.model.BotCommand;
 import com.pengrad.telegrambot.model.Update;
 import com.pengrad.telegrambot.request.SendMessage;
-
 import java.util.Objects;
 
 public interface Command {
