+++ bot/src/main/java/edu/java/bot/messageProcessors/MessageProcessor.java
@@ -1,14 +1,18 @@
-package edu.java.bot;
+package edu.java.bot.messageProcessors;
 
 import com.pengrad.telegrambot.model.Update;
 import com.pengrad.telegrambot.request.SendMessage;
-import edu.java.bot.commands.*;
+import edu.java.bot.commands.Command;
+import edu.java.bot.commands.HelpCommand;
+import edu.java.bot.commands.ListCommand;
+import edu.java.bot.commands.StartCommand;
+import edu.java.bot.commands.TrackCommand;
+import edu.java.bot.commands.UntrackCommand;
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
-
-import java.util.List;
 
 @Component
 @RequiredArgsConstructor
@@ -37,10 +41,11 @@ public class MessageProcessor implements UserMessageProcessor {
                 .stream()
                 .filter(command -> command.supports(update))
                 .toList();
+
         if (rightCommand.size() != 1) {
             return new SendMessage(
                     update.message().chat().id(),
-                    "Please enter a right command or choose from menu"
+                    "Please enter a right command or complete the input for the previous command"
             );
         } else {
             return rightCommand.getFirst().handle(update);
