+++ bot/src/main/java/edu/java/bot/messageProcessors/UserMessageProcessor.java
@@ -1,8 +1,8 @@
-package edu.java.bot;
+package edu.java.bot.messageProcessors;
 
 import com.pengrad.telegrambot.model.Update;
 import com.pengrad.telegrambot.request.SendMessage;
-
+import edu.java.bot.commands.Command;
 import java.util.List;
 
 public interface UserMessageProcessor {
