+++ bot/src/main/java/edu/java/bot/repositories/LinkRepository.java
@@ -1,11 +1,9 @@
 package edu.java.bot.repositories;
 
-
 import edu.java.bot.models.Link;
+import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
 
 @Repository
 public interface LinkRepository extends JpaRepository<Link, Long> {
