+++ bot/src/main/java/edu/java/bot/repositories/UserRepository.java
@@ -1,10 +1,9 @@
 package edu.java.bot.repositories;
 
 import edu.java.bot.models.User;
+import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
