+++ bot/src/main/java/edu/java/bot/services/LinkService.java
@@ -2,11 +2,10 @@ package edu.java.bot.services;
 
 import edu.java.bot.models.Link;
 import edu.java.bot.repositories.LinkRepository;
+import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
