+++ bot/src/test/java/edu/java/bot/commands/ListCommandTest.java
@@ -1,24 +1,22 @@
-package commands;
+package edu.java.bot.commands;
 
 import com.pengrad.telegrambot.model.Chat;
 import com.pengrad.telegrambot.model.Message;
 import com.pengrad.telegrambot.model.Update;
-import edu.java.bot.commands.TrackCommand;
 import edu.java.bot.models.Link;
 import edu.java.bot.models.User;
-import edu.java.bot.models.UserState;
-import edu.java.bot.services.LinkService;
 import edu.java.bot.services.UserService;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Optional;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 
-import java.util.Optional;
+public class ListCommandTest {
 
-public class TrackCommandTest {
     private static UserService userService;
-    private static LinkService linkService;
     private static Update update;
     private static Message message;
     private static Chat chat;
@@ -26,63 +24,66 @@ public class TrackCommandTest {
     @BeforeAll
     public static void mockInit() {
         userService = Mockito.mock(UserService.class);
-        linkService = Mockito.mock(LinkService.class);
         update = Mockito.mock(Update.class);
         message = Mockito.mock(Message.class);
         chat = Mockito.mock(Chat.class);
     }
 
     @Test
-    public void notRegisteredOrNeutralUser() {
+    public void notRegisteredUser() {
         Mockito.when(update.message()).thenReturn(message);
-        Mockito.when(message.text()).thenReturn("/track");
+        Mockito.when(message.text()).thenReturn("/list");
         Mockito.when(message.chat()).thenReturn(chat);
         Mockito.when(chat.id()).thenReturn(1L);
         Mockito.when(userService.findByChatId(Mockito.anyLong()))
                 .thenReturn(Optional.empty());
-        TrackCommand trackCommand = new TrackCommand(userService, linkService);
+        ListCommand listCommand = new ListCommand(userService);
         Assertions.assertEquals(
-                trackCommand.handle(update).getParameters().get("text"),
+                listCommand.handle(update).getParameters().get("text"),
                 "Please register! Use command /start"
-        );
-        Mockito.when(userService.findByChatId(Mockito.anyLong()))
-                .thenReturn(Optional.of(new User()));
-        Assertions.assertEquals(
-                trackCommand.handle(update).getParameters().get("text"),
-                "Please, enter the link you want to track"
         );
     }
 
     @Test
-    public void userWithTrackState() {
+    public void userTrackedLinks() {
         Mockito.when(update.message()).thenReturn(message);
-        Mockito.when(message.text()).thenReturn("www.example.com");
+        Mockito.when(message.text()).thenReturn("/list");
         Mockito.when(message.chat()).thenReturn(chat);
         Mockito.when(chat.id()).thenReturn(1L);
-        Link link = new Link();
-        link.setUrl("www.example.com");
-        Mockito.when(linkService.findByUrl("www.example.com")).thenReturn(Optional.of(link));
         User user = new User();
-        user.setState(UserState.TRACK);
+        user.setLinks(new ArrayList<>());
         Mockito.when(userService.findByChatId(Mockito.anyLong()))
                 .thenReturn(Optional.of(user));
-        TrackCommand trackCommand = new TrackCommand(userService, linkService);
+        ListCommand listCommand = new ListCommand(userService);
         Assertions.assertEquals(
-                trackCommand.handle(update).getParameters().get("text"),
-                "Link is now tracked!"
+                listCommand.handle(update).getParameters().get("text"),
+                "There are no tracked links"
+        );
+
+        Link link1 = new Link();
+        Link link2 = new Link();
+        link1.setUrl("www.example1.com");
+        link2.setUrl("www.example2.com");
+        user.setLinks(new ArrayList<>(List.of(link1, link2)));
+        Assertions.assertEquals(
+                listCommand.handle(update).getParameters().get("text"),
+                """
+                        Tracked Links:\s
+                        www.example1.com
+                        www.example2.com"""
         );
     }
 
     @Test
     public void commandAndDescription() {
-        TrackCommand trackCommand = new TrackCommand(userService, linkService);
+        ListCommand listCommand = new ListCommand(userService);
         Assertions.assertEquals(
-                trackCommand.command(),
-                "/track"
+                listCommand.command(),
+                "/list"
         );
         Assertions.assertEquals(
-                trackCommand.description(),
-                "start tracking link"
+                listCommand.description(),
+                "show list of tracked links"
         );
     }
 }
