+++ bot/src/test/java/edu/java/bot/commands/StartCommandTest.java
@@ -1,25 +1,25 @@
-package commands;
+package edu.java.bot.commands;
+
 
 import com.pengrad.telegrambot.model.Chat;
 import com.pengrad.telegrambot.model.Message;
 import com.pengrad.telegrambot.model.Update;
-import edu.java.bot.commands.StartCommand;
 import edu.java.bot.models.User;
 import edu.java.bot.services.UserService;
+import java.util.Optional;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 
-import java.util.Optional;
-
 public class StartCommandTest {
+
+
     @Test
     public void registrationUser() {
         UserService userService = Mockito.mock(UserService.class);
         Update update = Mockito.mock(Update.class);
         Message message = Mockito.mock(Message.class);
         Chat chat = Mockito.mock(Chat.class);
-
         Mockito.when(update.message()).thenReturn(message);
         Mockito.when(message.text()).thenReturn("/start");
         Mockito.when(message.chat()).thenReturn(chat);
@@ -27,9 +27,8 @@ public class StartCommandTest {
         User user = new User();
         user.setId(1L);
         user.setChatID(1L);
-        Mockito.when(userService.findByChatId(Mockito.anyLong()))
-                .thenReturn(Optional.of(user));
         StartCommand startCommand = new StartCommand(userService);
+        Mockito.when(userService.findByChatId(Mockito.anyLong())).thenReturn(Optional.of(user));
         Assertions.assertEquals(
                 startCommand.handle(update).getParameters().get("text"),
                 "Please, enter your name"
